Adds HONCHO Procfile support
This is really handy when firing up the multiple components required to
run the service, such as the high/low queue workers.

This also enables a better discernation among job types - as the low
priority queue is delegated to the transcoder options, which will take a
long time to run. The high queue is intended to be rapid ops like
xferring files offsite, and running DB updates/maintenance