Adds a Makefile target `depends` for spotify deps
The spotify components rely on a few things, namely libspotify and
an application key which can be obtained from:

https://devaccount.spotify.com/my-account/keys/

This is not readily obvious that there are such a crazy amount of
dependencies. So on first time setup, it seems feesable to execute
a bit of config management to install these dependencies.

Additionally, the depends target also inits a virtualenv if you dont
have one, and then kicks off the routine to install dependencies