Inserted additional dependencies to the installer script.
Upon further investigation, the RQ workers do not properly shut down on
exiting honcho, this leads me to beleive the forking behavior of the
processes make it a non-ideal candidates for honcho management. This
will be reverted for now and reintroduced later when python-rq has been
patched to allow non-daemon mode workers.