Re-enabling Honcho w/ multi-queue support
- adds worker-queues.py to handle spinning up associated queue worker
- Sets bitlist to run from production.ini by default w/ honcho

With this change, there are some one-off cases where the workers can be
left hanging around if the main honcho process tanks and takes teh
website with it. As the workers fork to the background - process
management here is tricky. However this is 50% better than just calling
rq worker [queue] - as those leave behind workers 100% of the time.